+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Set the minimum supported Rust version (MSRV) to 1.87. This is required by the `std::io::ErrorKind` variants used to map launch failures to exit codes, and by the optional `which_problem` dependency which relies on `OsStr::display` (stabilized in 1.87). (https://github.com/schneems/fun_run/pull/25)
 - Add `fun_run::OutputWithName` extension trait to construct a synthetic `NamedOutput` without running a process. Useful for testing code that inspects a `NamedOutput`/`CmdError` (https://github.com/schneems/fun_run/pull/23)
 - Add `fun_run::ExitStatusFromCode` extension trait as an ergonomic construct to build an `ExitStatus` without bit shifting (https://github.com/schneems/fun_run/pull/23)
+- Improve `exit status` display for `CmdError`. On Unix, when a signal kills a process it does not have a status code `ExitStatus::code()` returns `None`. Previously we were defaulting to a value of `1` for that case. Now, we are replicating the behavior of bash by adding `128` to the signal number such that `SIGTERM` becomes `143`.
 
 ## 0.6.0